Make clippy happy

+++ src/utils.rs
@@ -3,7 +3,7 @@ use std::{ffi::CStr, mem, ptr};
 pub fn get_username_from_uid(uid: u32) -> Option<String> {
     unsafe {
         let amt = match libc::sysconf(libc::_SC_GETPW_R_SIZE_MAX) {
-            n if n < 0 => 512 as usize,
+            n if n < 0 => 512_usize,
             n => n as usize,
         };
         let mut buf = Vec::with_capacity(amt);
